Bugfix 2387: Don't list interfaces in ipsec config file.
The "interfaces=..." entry in the /etc/ipsec.conf file needs to list
the actual interfaces we are using only if the underlying kernel IPsec
support is provided by KLIPS.  In our case, we are using NETKEY, so
we don't need to list our interfaces there.  Not listing them makes
ipsec startup a bit more robust.