Bugfix 3284: Allow commit to succeed if local-ip is not configured.
Previously, the ipsec setup script would fail the commit if the IPv4
address given in the local-ip parameter was not configured on
one of the ipsec-interfaces at the time that the commit took place.
This causes problems for PPP interfaces that may be configured, but
down for operational reasons.  This change makes the setup script more
liberal.  It will allow the commit to complete, but issue a warning
the address, they can restart ipsec. PPP has logic to restart
ipsec when a link comes up.