Bugfix 8276: Vti not working on a 32-bit machine due to sign bit overload.
In a 32-bit machine the signed bit is at the 32'nd bit position.
Mark used by vti set's (0x9000-0000) that bit position.
Changed the api to use strtoul to read the data from the command line and
configure the mark.
Also, changed the vyatta-cfg-vpn package to be arch dependent since it now
has a binary that it generates.